refactor(workstation): extract theme-picker state slice from inkViewModel
inkViewModel.ts's LogInkState/LogInkAction/applyLogInkAction mix every
view's state in one 3.7k-line module. Pull the theme picker's fields,
action union, reducer, and fuzzy-filter helpers into their own module
(themePicker.ts) as the first per-surface slice, following the
commitCompose.ts pattern already in the codebase.

State fields stay flat on LogInkState rather than nested — 13 call
sites read them directly and nesting would churn all of them for no
behavior change. The overlay-exclusivity logic (closing help/palette
when the picker opens) stays in the composition root since it's a
cross-slice concern, not something the picker owns. inkViewModel.ts
re-exports the slice's public helpers so no importer needed to change.

Test coverage for the pure slice logic moves to themePicker.test.ts;
inkViewModel.test.ts keeps only the composition-root-specific case.

This is a first-slice extraction, not the full decomposition — the
remaining 15 views' state stays in inkViewModel.ts for follow-up PRs.

Fixes #1630.
