+++ src/workstation/runtime/inkViewModel.ts
@@ -2,7 +2,12 @@ import { GitLogCommitRow, GitLogRow, getCommitRows } from '../../commands/log/da
 import { hashesMatchAny } from '../../git/hashes'
 import type { RebasePlanRow, RebaseTodoAction } from '../../git/rebasePlanActions'
 import type { ConflictRegion } from '../../git/conflictRegionActions'
-import { getLogInkThemePresets, type LogInkThemePreset } from '../chrome/theme'
+import {
+    applyThemePickerAction,
+    createThemePickerState,
+    type ThemePickerAction,
+} from './themePicker'
+export { filterThemePresets, getThemePickerSelection, getThemePickerSelectionFor } from './themePicker'
 import {
     CommitComposeAction,
     CommitComposeState,
@@ -606,10 +611,10 @@ export type LogInkState = {
   paletteSelectedIndex: number
   paletteRecent: string[]
   /**
-   * Theme picker (`gC`) interaction state. `themePickerFilter` is the
-   * typed substring query; `themePickerIndex` is a cursor into the
-   * filtered preset list. While open, the workstation live-previews the
-   * cursored theme (wired in `app.ts`).
+   * Theme picker (`gC`) interaction state — owned by `./themePicker.ts`
+   * (#1630 first slice). Fields stay flat here rather than nested so the
+   * many read call sites don't churn; the action/reducer logic and the
+   * fuzzy-filter helpers live in that module.
    */
   showThemePicker: boolean
   themePickerFilter: string
@@ -1289,11 +1294,7 @@ export type LogInkAction =
   | { type: 'clearHelpFilter' }
   | { type: 'toggleViewKeys' }
   | { type: 'toggleCommandPalette' }
-  | { type: 'toggleThemePicker' }
-  | { type: 'moveThemePicker'; delta: number; presetCount: number }
-  | { type: 'appendThemePickerFilter'; value: string }
-  | { type: 'backspaceThemePickerFilter' }
-  | { type: 'clearThemePickerFilter' }
+  | ThemePickerAction
   | { type: 'openGitignorePicker'; file: string }
   | { type: 'closeGitignorePicker' }
   | { type: 'moveGitignorePicker'; delta: number; count: number }
@@ -2094,9 +2095,7 @@ export function createLogInkState(
     paletteFilter: '',
     paletteSelectedIndex: 0,
     paletteRecent: [],
-    showThemePicker: false,
-    themePickerFilter: '',
-    themePickerIndex: 0,
+    ...createThemePickerState(),
     commitCompose: createCommitComposeState(),
     diffPreviewOffset: 0,
     worktreeDiffOffset: 0,
@@ -3209,48 +3208,23 @@ export function applyLogInkAction(state: LogInkState, action: LogInkAction): Log
         pendingKey: undefined,
       }
     }
-    case 'toggleThemePicker': {
-      const opening = !state.showThemePicker
+    case 'toggleThemePicker':
       return {
         ...state,
-        showThemePicker: opening,
+        ...applyThemePickerAction(state, action),
         // Only one overlay at a time — close help / palette / view-keys on open.
         showHelp: false,
         showViewKeys: false,
         showCommandPalette: false,
-        themePickerFilter: '',
-        themePickerIndex: 0,
         pendingKey: undefined,
       }
-    }
     case 'moveThemePicker':
-      return {
-        ...state,
-        themePickerIndex: clampIndex(
-          state.themePickerIndex + action.delta,
-          action.presetCount
-        ),
-        pendingKey: undefined,
-      }
     case 'appendThemePickerFilter':
-      return {
-        ...state,
-        themePickerFilter: `${state.themePickerFilter}${action.value}`,
-        themePickerIndex: 0,
-        pendingKey: undefined,
-      }
     case 'backspaceThemePickerFilter':
-      return {
-        ...state,
-        themePickerFilter: state.themePickerFilter.slice(0, -1),
-        themePickerIndex: 0,
-        pendingKey: undefined,
-      }
     case 'clearThemePickerFilter':
       return {
         ...state,
-        themePickerFilter: '',
-        themePickerIndex: 0,
+        ...applyThemePickerAction(state, action),
         pendingKey: undefined,
       }
     case 'openGitignorePicker':
@@ -3697,71 +3671,6 @@ export function intentOpenComposeForFile(
   return { type: 'navigateOpenComposeForFile', fileIndex: idx }
 }
 
-/**
- * Fuzzy (subsequence) score for a preset id against a lowercase query.
- * Returns `null` when the query chars don't appear in order; otherwise a
- * score where contiguous runs, a start-of-string match, and matches right
- * after a `-` separator are rewarded — so `gl` ranks `gruvbox-light` /
- * `github-light` above incidental matches, and `tn` finds `tokyo-night`.
- */
-function fuzzyScoreThemePreset(preset: string, query: string): number | null {
-  const target = preset.toLowerCase()
-  let qi = 0
-  let score = 0
-  let lastMatch = -2
-  for (let i = 0; i < target.length && qi < query.length; i += 1) {
-    if (target[i] === query[qi]) {
-      score += 1
-      if (i === lastMatch + 1) score += 4 // contiguous run
-      if (i === 0) score += 8 // matches the very start
-      else if (target[i - 1] === '-') score += 4 // start of a word segment
-      lastMatch = i
-      qi += 1
-    }
-  }
-  return qi === query.length ? score : null
-}
-
-/**
- * Filter the full preset list by a fuzzy (subsequence) query, ranked best
- * match first (ties broken by catalog order). An empty query returns every
- * preset in catalog order. Shared by the theme picker overlay renderer, the
- * input handler (for cursor bounds), and the live-preview selector so all
- * three agree on the same filtered list.
- */
-export function filterThemePresets(filter: string): LogInkThemePreset[] {
-  const query = filter.trim().toLowerCase()
-  const all = getLogInkThemePresets()
-  if (!query) {
-    return all
-  }
-  return all
-    .map((preset, index) => ({ preset, index, score: fuzzyScoreThemePreset(preset, query) }))
-    .filter((entry): entry is { preset: LogInkThemePreset; index: number; score: number } => entry.score !== null)
-    .sort((a, b) => b.score - a.score || a.index - b.index)
-    .map((entry) => entry.preset)
-}
-
-/**
- * The preset currently under the theme-picker cursor (clamped to the
- * filtered list). `undefined` when the filter matches nothing.
- */
-export function getThemePickerSelection(state: LogInkState): LogInkThemePreset | undefined {
-  return getThemePickerSelectionFor(state.themePickerFilter, state.themePickerIndex)
-}
-
-/**
- * State-model-agnostic variant: the preset under the picker cursor for a
- * raw `filter` + `index`. Used by the workspace top-level surface, which
- * keeps its own state shape but shares the picker filtering.
- */
-export function getThemePickerSelectionFor(
-  filter: string,
-  index: number
-): LogInkThemePreset | undefined {
-  const filtered = filterThemePresets(filter)
-  if (filtered.length === 0) {
-    return undefined
-  }
-  return filtered[clampIndex(index, filtered.length)]
-}
+// Theme picker fuzzy-filter + selection helpers live in ./themePicker.ts
+// (#1630 first slice) and are re-exported above so existing importers
+// don't churn.
